Chatbot-backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-pinecone
sentence-transformers
#ctransformers==0.2.5
pymupdf
openai
langchain-openai
-e .